\name{braid_append_text}
\alias{braid_append_text}
\title{Appends braid text.}
\usage{braid_append_text(braid, text="", reset=FALSE)}
\description{Appends braid text.}
\details{The output from braid_write is stored in a character vector, for later dumping to file.}
\keyword{internal}
\arguments{\item{braid}{A braid object}}
