\name{braid_heading}
\alias{braid_heading}
\title{Creates braid heading in latex format and writes results to output file.}
\usage{braid_heading(braid, x, headinglevel="chapter", pagebreak=FALSE)}
\description{Creates braid heading in latex format and writes results to output file.}
\details{Creates braid heading}
\arguments{\item{braid}{A braid object}
\item{x}{A character vector}
\item{headinglevel}{Character vector corresponding to latex heading, e.g. Chapter, Section, etc.}
\item{pagebreak}{Forces a page break if TRUE}}
