\name{braid_inc_counter}
\alias{braid_inc_counter}
\title{Updates braid counter.}
\usage{braid_inc_counter(braid)}
\description{Updates braid counter.}
\details{Each braid object has a static counter that can be used to generated automatic filenames. This function increments the counter and returns the result.}
\keyword{internal}
\arguments{\item{braid}{A braid object}}
