\name{braid_write}
\alias{braid_write}
\title{Appends text to braid output file.}
\usage{braid_write(braid, x, suffix="default")}
\description{Appends text to braid output file.}
\details{Appends text to braid output file.}
\arguments{\item{braid}{A braid object}
\item{x}{A character vector of text}
\item{suffix}{A character vector to be appended to x, by default a line break}}
