\name{braidAppendPlot}
\alias{braidAppendPlot}
\title{Appends braid plot.}
\usage{
  braidAppendPlot(braid, plot = NULL, filename = "", width
  = 4, height = 3, reset = FALSE)
}
\arguments{
  \item{braid}{A braid object}
}
\description{
  The output from braidWrite is stored in a character
  vector, for later dumping to file.
}
\keyword{internal}

