\name{braidAppendText}
\alias{braidAppendText}
\title{Appends braid text.}
\usage{
  braidAppendText(braid, text = "", reset = FALSE)
}
\arguments{
  \item{braid}{A braid object}
}
\description{
  The output from braidWrite is stored in a character
  vector, for later dumping to file.
}
\keyword{internal}

