\name{braidHeading}
\alias{braidHeading}
\title{Creates braid heading in latex format and writes results to output file.}
\usage{
  braidHeading(braid, x, headinglevel = "chapter",
    pagebreak = FALSE)
}
\arguments{
  \item{braid}{A braid object}

  \item{x}{A character vector}

  \item{headinglevel}{Character vector corresponding to
  latex heading, e.g. Chapter, Section, etc.}

  \item{pagebreak}{Forces a page break if TRUE}
}
\description{
  Creates braid heading
}

