\name{braidIncCounter}
\alias{braidIncCounter}
\title{Updates braid counter.}
\usage{
  braidIncCounter(braid)
}
\arguments{
  \item{braid}{A braid object}
}
\description{
  Each braid object has a static counter that can be used
  to generated automatic filenames. This function
  increments the counter and returns the result.
}
\keyword{internal}

