\name{braidSave}
\alias{braidSave}
\title{Writes result to braid output file.}
\usage{
<<<<<<< HEAD
  braidSave(b)
}
\arguments{
  \item{b}{object}
=======
  braidSave(braid)
}
\arguments{
  \item{braid}{A braid object}
>>>>>>> 03a698d53a139e14f9fda73084583835cb609986
}
\description{
  Writes result to braid output file.
}

