\name{braidWrite}
\alias{braidWrite}
\title{Appends text to braid output file.}
\usage{
  braidWrite(braid, x, suffix = "default")
}
\arguments{
  \item{braid}{A braid object}

  \item{x}{A character vector of text}

  \item{suffix}{A character vector to be appended to x, by
  default a line break}
}
\description{
  Appends text to braid output file.
}

